requirements.txt
transformer
datasets
accelerate
huggingfacehub